cinder-csi: Create storageclasses by default
Create a default storage class for cinder-csi with an optional volume type option.
The volume type can be specified with the cinder_csi_volume_type label. Fall
back to using the default cinder type if no volume type is defined. Additionally
enable the CluterRole access to list and get secrets for snapshot support.